// src/shared/api/generated/model/core/sessionsViewDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * SnapFlow API
 * REST API для SnapFlow
 * OpenAPI spec version: 1.0.0
 */

export interface SessionsViewDto {
  /** Id сессии устройства */
  deviceId: string;
  /** Название браузера */
  browserName: string;
  /** Версия браузера */
  browserVersion: string;
  /** Название операционной системы */
  osName: string;
  /** Версия операционной системы */
  osVersion: string;
  /** Имя устройства. Получаем из header "user-agent" */
  deviceName: string;
  /** Тип устройства */
  deviceType: string;
  /** Ip адрес устройства */
  ip: string;
  /** Дата последней генерации токенов */
  lastActive: string;
  /** Текущая сессия, с которой выполнен запрос */
  isCurrent: boolean;
}